Pull formatters into a separate module
